+++ app/src/main/java/com/example/pandia/luxury/interfaces/Writable.java
@@ -3,6 +3,6 @@ package com.example.pandia.luxury.interfaces;
 public interface Writable<T> {
     public void initializeWriter();
     public void writeEntry(T entry);
-    public int writtenEntries();
+    public long writtenEntries();
     public void finishWriter();
 }
